+++ tests/run.ts
@@ -27,6 +27,7 @@ async function main(): Promise<void> {
   await import('./textures.test');
   await import('./tutorial.test');
   await import('./friends.test');
+  await import('./invite.test');
   await import('./ads.test');
   await import('./icons.test');
   await import('./appIcon.test');